requirements/base.txt
psycopg2==2.5.4
Pillow==2.5.3
Django==1.7.1
djangorestframework==3.1.0
django_compressor==1.4
# Required by django-compressor
django-appconf==0.6
six==1.6.1
git+https://github.com/jbradberry/ultimate_tictactoe#egg=t3

BeautifulSoup4==4.3.2
